fix: extract component & update style properties
